+++ release_notes.md
@@ -4,16 +4,31 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker (metapackage) <version>
+### Microsoft.Azure.Functions.Worker (metapackage) 2.0.0
+
+- Updating `Microsoft.Azure.Functions.Worker.Core` to 2.0.0
+- Updating `Microsoft.Azure.Functions.Worker.Grpc` to 2.0.0
+
+### Microsoft.Azure.Functions.Worker.Core 2.0.0
 
 - `ValidateScopes` is enabled for developement environments by default. (#2705)
   - The default is the value of `IsDevelopment(IHostingEnvironment)`.
+- Capability `IncludeEmptyEntriesInMessagePayload` is now enabled by default (#2701)
+  - This means that empty entries will be included in the function trigger message payload by default.
+  - To disable this capability and return to the old behaviour, set `IncludeEmptyEntriesInMessagePayload` to `false` in the worker options.
 
-### Microsoft.Azure.Functions.Worker.Core <version>
+    ```csharp
+    var host = new HostBuilder()
+    .ConfigureFunctionsWorkerDefaults(builder =>
+    {
+    }, options =>
+    {
+        options.IncludeEmptyEntriesInMessagePayload = false;
+    })
+    .Build();
+    ```
 
-- <entry>
-
-### Microsoft.Azure.Functions.Worker.Grpc <version>
+### Microsoft.Azure.Functions.Worker.Grpc 2.0.0
 
 - Removed fallback command line argument reading code for grpc worker startup options. (#1908)